Remove patch from test samples
